pyproject.toml
[tool.poetry]
name = "sports-news"
version = "0.1.0"
description = ""
authors = ["Denis Mazurak <pure.demicube@gmail.com>"]
readme = "README.md"
packages = [{include = "sports_news"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-import-export = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
